+++ src/Web/OAuth/Internal.hs
@@ -14,6 +14,10 @@ module Web.OAuth.Internal
   , TokenResponse (TokenResponse, access_token, token_type, expires_in, refresh_token_resp)
   , TokenResponseHeaders
   , handleRegister
+  , registerServer
+  , handleRegistrationGet
+  , handleRegistrationUpdate
+  , handleRegistrationDelete
   , ClientRegistrationRequest (..)
   , ClientRegistrationResponse (..)
   , handleMetadata
@@ -23,7 +27,15 @@ module Web.OAuth.Internal
 import Web.OAuth.AuthorizeAPI (validateScope)
 import Web.OAuth.AuthorizeCallbackAPI (LoginForm (..), handleAuthorizeCallback)
 import Web.OAuth.MetadataAPI (OAuthMetadata (..), handleMetadata)
-import Web.OAuth.RegisterAPI (ClientRegistrationRequest (..), ClientRegistrationResponse (..), handleRegister)
+import Web.OAuth.RegisterAPI
+  ( ClientRegistrationRequest (..)
+  , ClientRegistrationResponse (..)
+  , handleRegister
+  , handleRegistrationDelete
+  , handleRegistrationGet
+  , handleRegistrationUpdate
+  , registerServer
+  )
 import Web.OAuth.TokenAPI
   ( TokenRequest (..)
   , TokenResponse (TokenResponse, access_token, token_type, expires_in, refresh_token_resp)